refactor: split tests into domain-specific files. Rewrite to use actual tool logic instead of replicating it
