+++ test/tools/helpers.js
@@ -1,0 +1,159 @@
+/**
+ * Shared test infrastructure for tool-level tests.
+ *
+ * createMockServer() — returns a minimal MCP server stub that captures
+ *   registerTool() calls, plus the resulting handlers map.
+ *
+ * MockAnyListClient — in-memory client that owns its own state arrays.
+ *   Call client.reset() (or create a fresh instance) in beforeEach.
+ */
+
+export function createMockServer() {
+  const handlers = {};
+  const server = {
+    registerTool: (name, _schema, handler) => { handlers[name] = handler; },
+    // elicitation.js calls server.server.getClientCapabilities() to detect support.
+    // Returning null means elicitation is disabled; missing-param paths throw instead.
+    server: { getClientCapabilities: () => null },
+  };
+  return { server, handlers };
+}
+
+export class MockAnyListClient {
+  constructor() {
+    this.client = null;
+    this.targetList = null;
+    this._connected = false;
+    this.defaultListName = null;
+    this._items = [];
+    this._lists = [];
+    this._favorites = [];
+    this._recents = [];
+    this._recipes = [];
+    this._events = [];
+    this._labels = [];
+    this._collections = [];
+    this._pendingImport = null;
+  }
+
+  reset() {
+    this.client = null;
+    this.targetList = null;
+    this._connected = false;
+    this._items = [];
+    this._lists = [];
+    this._favorites = [];
+    this._recents = [];
+    this._recipes = [];
+    this._events = [];
+    this._labels = [];
+    this._collections = [];
+    this._pendingImport = null;
+  }
+
+  async connect(listName = null) {
+    const name = listName || process.env.ANYLIST_LIST_NAME || 'Groceries';
+    this._connected = true;
+    const items = this._items;
+    this.targetList = {
+      name,
+      identifier: 'list-123',
+      items,
+      getItemByName: (n) => items.find(i => i.name.toLowerCase() === n.toLowerCase()) || null,
+    };
+    this.client = {};
+    return true;
+  }
+
+  getLists() { return this._lists; }
+
+  async addItem(name, qty, notes) {
+    this._items.push({ name, quantity: qty, notes });
+  }
+
+  async removeItem(name) {
+    const idx = this._items.findIndex(i => i.name === name);
+    if (idx === -1) throw new Error(`Item "${name}" not found in list, so can't check it`);
+    this._items[idx].checked = true;
+  }
+
+  async deleteItem(name) {
+    const idx = this._items.findIndex(i => i.name === name);
+    if (idx === -1) throw new Error(`Item "${name}" not found in list, so can't delete it`);
+    this._items.splice(idx, 1);
+  }
+
+  async getItems(includeChecked = false, includeNotes = false) {
+    let items = [...this._items];
+    if (!includeChecked) items = items.filter(i => !i.checked);
+    return items.map(i => ({
+      name: i.name,
+      quantity: i.quantity || 1,
+      checked: i.checked || false,
+      category: i.category || 'other',
+      ...(includeNotes && i.notes ? { note: i.notes } : {}),
+    }));
+  }
+
+  async getFavoriteItems() { return this._favorites; }
+  async getRecentItems() { return this._recents; }
+
+  async getRecipes(search = null) {
+    let r = [...this._recipes];
+    if (search) r = r.filter(x => x.name.toLowerCase().includes(search.toLowerCase()));
+    return r;
+  }
+
+  async getRecipeDetails(name) {
+    const r = this._recipes.find(x => x.name.toLowerCase() === name.toLowerCase());
+    if (!r) throw new Error(`Recipe "${name}" not found`);
+    return { ...r, ingredients: r.ingredients || [], preparationSteps: r.preparationSteps || [] };
+  }
+
+  async createRecipe(opts) {
+    this._recipes.push(opts);
+    return { identifier: 'r-1', name: opts.name };
+  }
+
+  async deleteRecipe(name) {
+    const idx = this._recipes.findIndex(r => r.name.toLowerCase() === name.toLowerCase());
+    if (idx === -1) throw new Error(`Recipe "${name}" not found`);
+    this._recipes.splice(idx, 1);
+  }
+
+  async importRecipeFromUrl(url) {
+    if (!this._pendingImport) throw new Error('Could not parse recipe from URL. The site may not be supported.');
+    const imp = this._pendingImport;
+    this._recipes.push({ identifier: 'r-imported', name: imp.name, ...imp });
+    return {
+      name: imp.name,
+      identifier: 'r-imported',
+      ingredientCount: imp.ingredientCount || 0,
+      stepCount: imp.stepCount || 0,
+      source: imp.source || null,
+      sourceUrl: imp.sourceUrl || url,
+    };
+  }
+
+  async getMealPlanEvents() { return [...this._events]; }
+  async getMealPlanLabels() { return [...this._labels]; }
+
+  async createMealPlanEvent(opts) {
+    this._events.push(opts);
+    return { identifier: 'e-1', date: opts.date };
+  }
+
+  async deleteMealPlanEvent(id) {
+    const idx = this._events.findIndex(e => e.identifier === id);
+    if (idx === -1) throw new Error(`Meal plan event "${id}" not found`);
+    this._events.splice(idx, 1);
+  }
+
+  async getRecipeCollections() { return [...this._collections]; }
+
+  async createRecipeCollection(name, recipeNames = []) {
+    const c = { identifier: 'c-1', name, recipeCount: recipeNames.length, recipeNames };
+    this._collections.push(c);
+    return c;
+  }
+}
