+++ test/tools/shopping.test.js
@@ -1,0 +1,139 @@
+import { describe, it, beforeEach } from 'node:test';
+import assert from 'node:assert/strict';
+import { register } from '../../src/tools/shopping.js';
+import { MockAnyListClient, createMockServer } from './helpers.js';
+
+describe('shopping tool', () => {
+  let client;
+  let handlers;
+
+  beforeEach(() => {
+    client = new MockAnyListClient();
+    const { server, handlers: h } = createMockServer();
+    register(server, () => Promise.resolve(client));
+    handlers = h;
+  });
+
+  describe('add_item', () => {
+    it('adds an item', async () => {
+      const result = await handlers.shopping({ action: 'add_item', name: 'Milk' });
+      assert.ok(result.content[0].text.includes('Successfully added "Milk"'));
+      assert.equal(client._items.length, 1);
+      assert.equal(client._items[0].name, 'Milk');
+    });
+
+    it('adds item with quantity and notes', async () => {
+      await handlers.shopping({ action: 'add_item', name: 'Eggs', quantity: 2, notes: 'organic' });
+      assert.equal(client._items[0].quantity, 2);
+      assert.equal(client._items[0].notes, 'organic');
+    });
+  });
+
+  describe('check_item', () => {
+    it('checks off an existing item', async () => {
+      client._items.push({ name: 'Milk', checked: false });
+      const result = await handlers.shopping({ action: 'check_item', name: 'Milk' });
+      assert.ok(result.content[0].text.includes('Successfully checked off'));
+      assert.equal(client._items[0].checked, true);
+    });
+
+    it('returns error for non-existent item', async () => {
+      const result = await handlers.shopping({ action: 'check_item', name: 'Nonexistent' });
+      assert.equal(result.isError, true);
+      assert.ok(result.content[0].text.includes('not found'));
+    });
+  });
+
+  describe('delete_item', () => {
+    it('deletes an existing item', async () => {
+      client._items.push({ name: 'Milk' });
+      const result = await handlers.shopping({ action: 'delete_item', name: 'Milk' });
+      assert.ok(result.content[0].text.includes('Successfully deleted'));
+      assert.equal(client._items.length, 0);
+    });
+
+    it('returns error for non-existent item', async () => {
+      const result = await handlers.shopping({ action: 'delete_item', name: 'Ghost' });
+      assert.equal(result.isError, true);
+    });
+  });
+
+  describe('list_items', () => {
+    it('returns empty message when no items', async () => {
+      const result = await handlers.shopping({ action: 'list_items' });
+      assert.ok(result.content[0].text.includes('No unchecked items'));
+    });
+
+    it('lists items grouped by category', async () => {
+      client._items.push({ name: 'Milk', category: 'Dairy' }, { name: 'Bread', category: 'Bakery' });
+      const result = await handlers.shopping({ action: 'list_items' });
+      assert.ok(result.content[0].text.includes('Milk'));
+      assert.ok(result.content[0].text.includes('Bread'));
+      assert.ok(result.content[0].text.includes('Dairy'));
+      assert.ok(result.content[0].text.includes('Bakery'));
+    });
+
+    it('excludes checked items by default', async () => {
+      client._items.push({ name: 'Milk', checked: false }, { name: 'Done', checked: true });
+      const result = await handlers.shopping({ action: 'list_items' });
+      assert.ok(result.content[0].text.includes('Milk'));
+      assert.ok(!result.content[0].text.includes('Done'));
+    });
+
+    it('includes checked items when requested', async () => {
+      client._items.push({ name: 'Milk', checked: false }, { name: 'Done', checked: true });
+      const result = await handlers.shopping({ action: 'list_items', include_checked: true });
+      assert.ok(result.content[0].text.includes('Done'));
+    });
+
+    it('includes notes when requested', async () => {
+      client._items.push({ name: 'Milk', notes: 'whole milk' });
+      const result = await handlers.shopping({ action: 'list_items', include_notes: true });
+      assert.ok(result.content[0].text.includes('whole milk'));
+    });
+  });
+
+  describe('list_lists', () => {
+    it('returns empty message when no lists', async () => {
+      const result = await handlers.shopping({ action: 'list_lists' });
+      assert.ok(result.content[0].text.includes('No lists found'));
+    });
+
+    it('returns list names with counts', async () => {
+      client._lists = [
+        { name: 'Groceries', uncheckedCount: 5 },
+        { name: 'Costco', uncheckedCount: 2 },
+      ];
+      const result = await handlers.shopping({ action: 'list_lists' });
+      assert.ok(result.content[0].text.includes('Groceries'));
+      assert.ok(result.content[0].text.includes('5 unchecked'));
+    });
+  });
+
+  describe('get_favorites', () => {
+    it('returns empty message when no favorites', async () => {
+      const result = await handlers.shopping({ action: 'get_favorites' });
+      assert.ok(result.content[0].text.includes('No favorite items'));
+    });
+
+    it('returns favorite items', async () => {
+      client._favorites = [{ name: 'Bananas', details: 'organic' }];
+      const result = await handlers.shopping({ action: 'get_favorites' });
+      assert.ok(result.content[0].text.includes('Bananas'));
+      assert.ok(result.content[0].text.includes('organic'));
+    });
+  });
+
+  describe('get_recents', () => {
+    it('returns empty message when no recents', async () => {
+      const result = await handlers.shopping({ action: 'get_recents' });
+      assert.ok(result.content[0].text.includes('No recent items'));
+    });
+
+    it('returns recent items', async () => {
+      client._recents = [{ name: 'Avocado' }];
+      const result = await handlers.shopping({ action: 'get_recents' });
+      assert.ok(result.content[0].text.includes('Avocado'));
+    });
+  });
+});
